src/main/java/ru/scndjk/dsa/Stack/ParenthesesBalancer.java
package ru.scndjk.dsa.Stack;

public class ParenthesesBalancer {
    public static boolean isParenthesesBalanced(String expression) {
        Stack<Character> stack = new Stack<>();

        for (char c : expression.toCharArray()) {
            if (c == '(') {
                stack.push(c);
            } else if (c == ')') {
                if (stack.isEmpty() || stack.pop() != '(') {
                    return false;
                }
            }
        }

        return stack.isEmpty();
    }
}
